Refactor IL metadata operand resolution
Move the private metadata identity resolver and signature formatter into a focused IlBodyDiff partial without changing its surface or behavior.
